fix: rename keyword arguments in generator
The fields in the argument struct were being postfixed with `_` but their use
as variables in arguments were not.

Fixes #999
