+++ cynic-querygen/tests/queries/misc/keyword-argument.graphql
@@ -1,0 +1,7 @@
+# A nonsense name for a UUID field, but it's good enough for the test
+query MyQuery($type: UUID!) {
+  bar(id: $type) {
+    id
+    name
+  }
+}
